pyproject.toml
[tool.poetry]
name = "negetis"
version = "0.1.3"
description = "Static site generator"
authors = ["Dmitry Vysochin <dmitry.vysochin@gmail.com>"]
license = "MIT"
homepage = "http://negetis.com"
repository = "https://github.com/AxGrid/neGetiS"
keywords = ["static", "generator", "web", "html", "markdown"]
readme = "README.md"
classifiers = [
    "Topic :: Software Development :: Build Tools"
]

packages = [
    { include = "negetis" },
]
include = [
	"negetis/resources"
]


[tool.poetry.dependencies]
python = "^3.5"
Click = "^7.0"
jinja2 = "^2.10"
python-i18n = {version = "^0.3.7", extras = ["YAML"]}
pyyaml = "^5.1"
werkzeug = "^0.16.0"
pytz = "^2019.3"
deepmerge = "^0.1.0"
pyyaml-include = "^1.1"
markdown = "^3.1"
asq = "^1.3"
watchdog = "^0.9.0"
deprecation = "^2.0"
python-thumbnails = "^0.5.1"
pillow = "^6.2"


[tool.poetry.scripts]
negetis = 'negetis.main:run'

[tool.poetry.dev-dependencies]
pytest = "^5.2"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"


